Add retry on failed records in Kinesis sink
This adds an option to allow for the sink to retry sending any
events after receiving an error from Kinesis.  The sink
will inspect the response from Kinesis and only resend the
records that failed up to the max number of attempts.

Retrys are performed within the same Flume transaction.  Once
the max attempts count is reached Flume will commit the transaction
and discard any remaining failed events.  Thus delivery
is not guaranteed (although we can address in the future by
adding an option for unlimited retries).